+++ static/svgdx-editor.js
@@ -264,6 +264,73 @@ const textViewer = CodeMirror(document.getElementById('text-output'), {
         svg.setAttribute('viewBox', saved_viewbox);
     });
 
+    // copy PNG button
+    document.getElementById('copy-png').addEventListener('click', async () => {
+        await copyPng();
+    }, false);
+
+    async function copyPng() {
+        const svg = document.querySelector('#svg-output svg');
+        const saved_viewbox = svg.getAttribute('viewBox');
+        // temporarily set width, height, and viewBox to original values
+        svg.setAttribute('width', original_width);
+        svg.setAttribute('height', original_height);
+        svg.setAttribute('viewBox', original_viewbox);
+
+        // scale to a consistent (high) resolution
+        // TODO: additionally support a lower (e.g. 512px) resolution for smaller PNGs
+        // probably from a hover menu on the button
+        const maxDim = 2048;
+        let pxWidth = svg.width.baseVal.value;
+        let pxHeight = svg.height.baseVal.value;
+        if (pxWidth > pxHeight) {
+            pxHeight = (maxDim / pxWidth) * pxHeight;
+            pxWidth = maxDim;
+        } else {
+            pxWidth = (maxDim / pxHeight) * pxWidth;
+            pxHeight = maxDim;
+        }
+        svg.setAttribute('width', pxWidth);
+        svg.setAttribute('height', pxHeight);
+
+        const img = new Image();
+        img.src = URL.createObjectURL(new Blob([svg.outerHTML], { type: "image/svg+xml" }));
+        img.width = pxWidth;
+        img.height = pxHeight;
+
+        await new Promise((resolve) => {
+            img.onload = resolve;
+        });
+
+        const canvas = document.createElement("canvas");
+        const context = canvas.getContext("2d");
+        canvas.width = img.width;
+        canvas.height = img.height;
+        context.drawImage(img, 0, 0);
+
+        // Release the object URL now it's in the canvas
+        URL.revokeObjectURL(img.src);
+
+        const pngBlob = await new Promise((resolve) => {
+            canvas.toBlob((blob) => resolve(blob), "image/png");
+        });
+        try {
+            await navigator.clipboard.write([
+              new ClipboardItem({
+                [pngBlob.type]: pngBlob,
+              }),
+            ]);
+            console.log("PNG image copied to clipboard!");
+        } catch (error) {
+            console.error("Error copying PNG image to clipboard:", error);
+        }
+
+        // restore previous values
+        svg.setAttribute('width', '100%');
+        svg.setAttribute('height', '100%');
+        svg.setAttribute('viewBox', saved_viewbox);
+    }
+
     // copy as base64 button
     document.getElementById('copy-base64').addEventListener('click', () => {
         const svg = document.querySelector('#svg-output svg');